Update readme information and set migration strategy with fails

+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/com/peryloth/r2dbc/entity/UsuarioEntity.java
@@ -3,16 +3,17 @@ package com.peryloth.r2dbc.entity;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Table;
 
+import java.math.BigInteger;
 import java.time.LocalDate;
 
 @Table("usuario")
 public class UsuarioEntity {
     @Id
-    private Long id;
+    private BigInteger id;
     private String nombre;
     private String apellido;
     private String email;
     private LocalDate fechaNacimiento;
     private String direccion;
-    private long salarioBase;
+    private Long salarioBase;
 }
